ai_tags/requirements.txt
mistralai==1.5.1
fastapi==0.115.11
uvicorn==0.34.0
requests==2.32.3
boto3==1.37.33
botocore==1.37.33
PyMuPDF==1.25.5
python-docx==1.1.2
pyahocorasick==2.1.0